+++ Sources/MapboxMaps/Annotations/ViewAnnotationManager.swift
@@ -60,6 +60,13 @@ public final class ViewAnnotationManager {
     /// The default value is true, and setting this value to false will disable the validation.
     public var validatesViews = true
 
+    /// The complete list of annotations associated with the receiver.
+    public var annotations: [UIView: ViewAnnotationOptions] {
+        idsByView.compactMapValues { [mapboxMap] id in
+            try? mapboxMap.options(forViewAnnotationWithId: id)
+        }
+    }
+
     internal init(containerView: UIView, mapboxMap: MapboxMapProtocol) {
         self.containerView = containerView
         self.mapboxMap = mapboxMap
@@ -374,7 +381,6 @@ extension ViewAnnotationManager: DelegatingViewAnnotationPositionsUpdateListener
     internal func onViewAnnotationPositionsUpdate(forPositions positions: [ViewAnnotationPositionDescriptor]) {
         placeAnnotations(positions: positions)
     }
-
 }
 
 private extension ViewAnnotationPositionDescriptor {